requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Scheduling
apscheduler>=3.10.0

# Data Validation and Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.25.0

# Environment and Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Web Framework Extensions
python-multipart>=0.0.6

# Logging
structlog>=23.2.0

# Timezone Support
pytz>=2023.3

# CLI Dependencies
click>=8.1.0
rich>=13.0.0

# Development Dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
# black>=23.0.0
# isort>=5.12.0
# flake8>=6.0.0
# mypy>=1.7.0

# Song Recommendation Dependencies
sentence-transformers>=2.2.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.0  # Use faiss-gpu for GPU support
